requirements/common.txt
fastapi==0.67.0
kaggle==1.5.12
orjson==3.6.0
pandas==1.3.2
pydantic==1.8.2
tortoise-orm[aiomysql]==0.17.6
uvicorn==0.14.0